backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
sentence-transformers==2.2.2
scikit-learn==1.3.2
pypdf==3.17.4
python-docx==1.1.0
pydantic==2.5.0



